hecathon practise website/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
geopandas>=0.14.1
shapely>=2.0.2
earthengine-api>=0.1.381
torch>=2.1.1
rasterio>=1.3.9
reportlab>=4.0.7
pydantic>=2.5.2
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.26.2
pyproj>=3.6.1
python-dotenv>=1.0.0
